Updated id type to long and implemented method in PlantsService

+++ src/back-end/PlantCare.Application/Plants/Interfaces/IPlantsService.cs
@@ -11,6 +11,6 @@ public interface IPlantsService
     public Task<Result<PlantDto?>> GetByIdAsync(long id);
     public Task<Result<CreateUpdatePlantResponse>> CreateAsync(CreatePlantRequest req);
     public Task<Result<CreateUpdatePlantResponse>> UpdateAsync(UpdatePlantRequest req);
-    public Task<Result> DeleteAsync(int id);
+    public Task<Result> DeleteAsync(long id);
     public Task<Result<List<PlantDto>>> GetAllByUserAsync(User user);
 }
